refactor: move other stores cargo form to separate page
